idea edit form
